refactor blank and grid components
