+++ templates/Web/Pages/React.Blank/src/components/ReactBlank/ReactBlank.jsx
