+++ templates/Web/Pages/React.Grid/src/components/ReactGrid/Item.jsx
@@ -2,18 +2,18 @@
 import ImgGreyBox from "../../images/GreyBox.svg"
 import PropTypes from "prop-types";
 
-const GridComponent = ({ gridItem }) => {
+const Item = ({ item }) => {
   return (
     <div className="col-md-4 col-sm-12 p-5">
       <img src={ImgGreyBox} alt="Default Grey Box" className="mb-3" />
-      <h3>{gridItem.header}</h3>
-      <p>{gridItem.shortDescription}</p>
+      <h3>{item.header}</h3>
+      <p>{item.shortDescription}</p>
     </div>
   );
 }
 
-GridComponent.propTypes = {
-  gridItem: PropTypes.any
+Item.propTypes = {
+  item: PropTypes.any
 }
 
-export default GridComponent;
+export default Item;
