+++ templates/Web/_composition/ReactJS/Page.React.AddConstGrid/src/components/wts.ItemName/wts.ItemName_postaction.jsx
@@ -1,0 +1,10 @@
+﻿import React, { useState } from "react";
+import classnames from "classnames";
+import WarningMessage from "../WarningMessage";
+import styles from "./styles.module.css";
+import CONSTANTS from "../../constants";
+import Item from "./Item";
+
+//{[{
+const wts.ItemName = () => {
+//}]}
