+++ templates/Web/_composition/ReactJS/Page.React.AddImport/src/App_postaction.jsx
@@ -5,7 +5,7 @@ import NavBar from "./components/NavBar";
 import Footer from "./components/Footer";
 //^^
 //{[{
-import wts.ItemName from "./components/wts.ItemName";
+import wts.ItemName from "./components/wts.ItemName/wts.ItemName";
 //}]}
 //TODO Web Template Studio: Add routes for your new pages here.
 const App = () => {
