+++ templates/Web/_composition/ReactJS/Page.React.AddReturnGrid/src/components/wts.ItemName/wts.ItemName_postaction.jsx
@@ -1,0 +1,7 @@
+﻿    );
+}
+
+//{[{
+export default wts.ItemName;
+//}]}
+//^^
